Move DTOs to reside with Controllers
